Reorganize to avoid accidental async import
